LayoutTests/fast/dom/HTMLElement/attr-missing-parent-false.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<div id="div" contentEditable="false">
    <p id="p"></p>
</div>
<script>
description('When contentEditable is missing, it maps to "inherit" state and the element is NOT editable if its parent is NOT.')

shouldBe('document.getElementById("p").hasAttribute("contentEditable")','false');
shouldBe('document.getElementById("p").contentEditable', '"inherit"');
shouldBe('document.getElementById("p").isContentEditable', 'false');
shouldBe('window.getComputedStyle(p, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

document.getElementById("div").style.display= 'none';
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
